# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotImplementedResponse": {
        "additionalProperties": false,
        "description": "Body returned by a route whose backing service is not built yet.\n\nThese routes are registered rather than omitted so that the OpenAPI document\nis the complete v1 surface: the frontend can be written against the whole API\nfrom day one and build only the parts that answer. The alternative \u2014\npublishing only the built half \u2014 means the contract changes under the\nfrontend halfway through the project, which is the failure mode a contract\nfreeze exists to prevent.\n\n`service` names what is missing and `phase` names when it lands, so nobody has\nto guess whether a 503 is a bug or a scheduled gap.",
        "properties": {
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "phase": {
            "description": "Plan phase that delivers it",
            "title": "Phase",
            "type": "string"
          },
          "reason": {
            "default": "This endpoint is part of the frozen v1 contract but its backing ",
            "title": "Reason",
            "type": "string"
          },
          "service": {
            "description": "Service module not yet written",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "default": 503,
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "endpoint",
          "service",
          "phase"
        ],
        "title": "NotImplementedResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Supervisory Analytics Tool for SOC Assessment (SAT-SA) for NCIIPC.\n\nOutputs are **prioritised hypotheses with evidence**, never verdicts or\ncompliance grades. `Not assessable` is a first-class outcome, distinct from\n\"low risk\".\n\nFully offline: no cloud APIs, no CDN, no generative AI in the decision path.\n\n## Reading this contract\n\n`Not assessable` is a **value**, not an absence. Every score field is nullable\nand `null` means \"could not be assessed\" \u2014 never \"scored zero\". The frontend must\nnot coalesce `null` to `0`; a dimension with no data and a dimension with a\nmeasured score of zero are different claims.\n\n`POST /verdicts` records **a supervisor's own conclusion**, written by a human.\nNo Orion-derived payload in this API contains a verdict, a compliance grade, or a\npass/fail.\n\nRoutes marked `503` are part of this frozen contract but their backing service is\nnot written yet. The response body names the service and the phase that delivers\nit. They are listed rather than omitted so the surface does not change shape\nmid-project; `GET /api/v1` splits them into `live` and `pending`.",
    "title": "Orion (SAT-SA) Core Backend",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trends": {
      "get": {
        "description": "503 until Phase 15 (2.15). A series is only meaningful alongside its assessability per period: an entity that stopped submitting data has a gap in its series, and that gap must not be drawn as a line through it.",
        "operationId": "trends_api_v1_trends_get",
        "parameters": [
          {
            "description": "Entity to trace",
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "default": "",
              "description": "Entity to trace",
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "description": "egi, nsi, dts or sap",
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "default": "egi",
              "description": "egi, nsi, dts or sap",
              "title": "Metric",
              "type": "string"
            }
          },
          {
            "description": "Inclusive start, YYYY-MM-DD",
            "in": "query",
            "name": "period_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start, YYYY-MM-DD",
              "title": "Period Start"
            }
          },
          {
            "description": "Inclusive end, YYYY-MM-DD",
            "in": "query",
            "name": "period_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive end, YYYY-MM-DD",
              "title": "Period End"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Not built yet"
          }
        },
        "summary": "A scored metric over successive periods",
        "tags": [
          "trends"
        ],
        "x-orion": {
          "phase": "Phase 15 (2.15)",
          "service": "trend_service",
          "status": "pending"
        }
      }
    },
    "/api/v1/trends/change-points": {
      "get": {
        "description": "503 until Phase 15 (2.15). A change point marks where the shape of the values changed across the cohort. It is a prompt to look, not a finding: a genuine improvement produces one exactly as reliably as a deterioration, and the response must be able to say which direction it moved.",
        "operationId": "change_points_api_v1_trends_change_points_get",
        "parameters": [
          {
            "description": "egi, nsi, dts or sap",
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "default": "egi",
              "description": "egi, nsi, dts or sap",
              "title": "Metric",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Not built yet"
          }
        },
        "summary": "Periods where a metric's distribution shifted",
        "tags": [
          "trends"
        ],
        "x-orion": {
          "phase": "Phase 15 (2.15)",
          "service": "trend_service",
          "status": "pending"
        }
      }
    },
...
  },
...
}